+++ composer.json
@@ -91,8 +91,10 @@
         "viserio/events": "self.version",
         "viserio/exception": "self.version",
         "viserio/filesystem": "self.version",
+        "viserio/foundation": "self.version",
         "viserio/http": "self.version",
         "viserio/http-factory": "self.version",
+        "viserio/http-foundation": "self.version",
         "viserio/log": "self.version",
         "viserio/mail": "self.version",
         "viserio/monolog-bridge": "self.version",
