+++ src/Viserio/Component/Foundation/composer.json
@@ -11,9 +11,8 @@
         "config",
         "log",
         "exception",
-        "events",
         "parser",
-        "routing"
+        "console"
     ],
     "homepage": "http://github.com/narrowspark/framework",
     "license": "MIT",
@@ -42,30 +41,22 @@
         "psr/http-message-implementation": "^1.0"
     },
     "require-dev": {
-        "cakephp/chronos": "^1.0.4",
         "mockery/mockery": "^1.2.0",
-        "narrowspark/http-status": "^4.1.0",
         "narrowspark/testing-helper": "^7.0.0",
         "paragonie/halite": "^4.4.2",
         "phpunit/phpunit": "^7.2.0",
-        "psr/http-factory": "^1.0.0",
         "symfony/filesystem": "^4.1.0",
-        "symfony/process": "^4.1.0",
         "viserio/config": "self.version",
-        "viserio/cron": "self.version",
-        "viserio/events": "self.version",
         "viserio/exception": "self.version",
-        "viserio/pipeline": "self.version"
+        "viserio/http-factory": "self.version",
+        "vlucas/phpdotenv": "^2.5.0"
     },
     "suggest": {
-        "cakephp/chronos": "Required to use CheckForMaintenanceModeMiddleware and DownCommand (^1.0.4)",
-        "narrowspark/http-status": "Required to use the Down-, UpCommand and MaintenanceModeException (^4.1.0).",
         "psr/http-factory": "Required to use the SetRequestForConsole (^1.0.0).",
         "viserio/config": "Required to use the LoadConfiguration bootstrap class (self.version).",
-        "viserio/pipeline": "Required to use psr-15 middleware (self.version).",
         "viserio/statical-proxy": "Required to use statical-proxy bootstrap class (self.version).",
         "viserio/exception": "Required for a better http and console exception output (self.version).",
-        "vlucas/phpdotenv": "Required to use the LoadEnvironmentVariables bootstrap class (^2.3.0)."
+        "vlucas/phpdotenv": "Required to use the LoadEnvironmentVariables bootstrap class (^2.5.0)."
     },
     "extra": {
         "branch-alias": {
