+++ src/Viserio/Component/HttpFoundation/.changelog
@@ -1,0 +1,13 @@
+<?php
+declare(strict_types=1);
+
+use ChangelogGenerator\ChangelogConfig;
+
+return [
+     new ChangelogConfig(
+        'narrowspark',
+        'framework',
+        'v' . \Viserio\Component\Foundation\AbstractKernel::VERSION,
+        ['Component: Foundation']
+    )
+];
