+++ src/Viserio/Component/HttpFoundation/.gitattributes
@@ -1,0 +1,5 @@
+Tests/               export-ignore
+phpunit.xml.dist     export-ignore
+.gitattributes       export-ignore
+.gitignore           export-ignore
+.changelog_generator export-ignore
