+++ src/Viserio/Component/HttpFoundation/Tests/Fixture/Middleware/framework/down
@@ -1,0 +1,5 @@
+{
+    "time": 1491737578,
+    "message": "test",
+    "retry": 1
+}
