+++ src/Viserio/Component/HttpFoundation/automatic.json
@@ -1,0 +1,28 @@
+{
+  "configurators": {
+    "providers": {
+      "Viserio\\Component\\Foundation\\Provider\\ConsoleCommandsServiceProvider": [
+        "global"
+      ],
+      "Viserio\\Component\\Foundation\\Provider\\FoundationDataCollectorServiceProvider": [
+        "local",
+        "testing"
+      ],
+      "Viserio\\Component\\Foundation\\Provider\\WebServerServiceProvider": [
+        "local",
+        "testing"
+      ]
+    }
+  },
+  "post-install-output": [
+    "",
+    "  * <fg=blue>Run</> your http application:",
+    "    1. Change to the project directory",
+    "    2. Execute the <comment>php -S 127.0.0.1:8000 -t public</> command",
+    "    3. Browse to the <comment>http://localhost:8000/</> URL.",
+    "",
+    "       Quit the server with CTRL-C.",
+    "       Run <comment>composer require viserio/web-server --dev</> for a better web server.",
+    "  * <fg=blue>Read</> the documentation at <comment>https://narrowspark.com/docs</>"
+  ]
+}
